% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSSMfs.R
\name{BSSMfs}
\alias{BSSMfs}
\title{Fit the Bayesian subject-specific bi-level feature selection model described in Baer et al. (2021)}
\usage{
BSSMfs(
  Y_transpose,
  X,
  Z,
  p_k,
  nsim,
  burn,
  thin,
  mcmc_div = floor((nsim - burn)/4),
  covariate_select = c(NA),
  d = nrow(Y_transpose),
  Q,
  nu_0 = ncol(Z[[1]]),
  C_0,
  alpha = 10^-3,
  gamma = 10^-3,
  o = 10^-3,
  u = 10^-3,
  a_1 = 1,
  b_1 = 1,
  g = 1,
  h = 1,
  raw_MCMC_output = F
)
}
\arguments{
\item{Y_transpose}{A \mjeqn{J \times n}{}
matrix of longitudinal outcome data,
where \mjeqn{J}{J}
 is the number of occasions,
and \mjeqn{n}{n} is the number of subjects.
This model assumes multivariate normality of these outcome data.}

\item{X}{An \mjeqn{n \times p}{}
matrix of time-invariant feature data, where
\mjeqn{p}{p} is the number of features.}

\item{Z}{A list consisting of
\mjeqn{n \underset{J\times q}{\mathbf{Z}_i}}{}
matrices, where each
\mjeqn{\underset{J\times q}{\mathbf{Z}_i}}{}
is the matrix
of (time-varying) subject-specific feature data
corresponding to the subject-specific/random effect parameters,
\mjeqn{\underset{q\times 1}{\mathop{{{\mathbf{b}}_{i}}}}\,}{}
for
\mjeqn{i = 1, \cdots, n}{}.}

\item{p_k}{A
\mjeqn{K \times 1}{}
matrix of the feature group sizes,
\mjeqn{p_k}{},
where
\mjeqn{K}{K}
is the number of feature groups.
Note that
\mjeqn{\sum\limits_{k=1}^{K}{{{p}_{k}}}=p}{}.}

\item{nsim}{The total number of MCMC
iterations to be performed.}

\item{burn}{The number of MCMC
iterations to be discarded as part of the burn-in period.}

\item{thin}{The thinning interval applied to the MCMC
sample.}

\item{mcmc_div}{The interval at which the MCMC
iteration number will be reported; the
default value is
\code{floor((nsim-burn)/4)}.}

\item{covariate_select}{A vector denoting which of the
\mjeqn{p}{p} features the user wishes to remove from
the feature selection process; note that setting
\code{covariate_select<-c(NA)} removes no features from the feature
selection process; \code{covariate_select<-c(NA)} is the default setting.}

\item{d}{The degrees of freedom hyper-prior for the inverse-Wishart-distributed
covariance matrix,
\mjeqn{\underset{J\times J}{\mathbf{\Sigma}}}{};
note that
\mjeqn{\underset{J\times J}{\mathbf{\Sigma}}}{}
represents the prior
covariance for each
of the
\mjeqn{l = 1, \cdots, p}{}
time-varying feature parameters
within the matrix
\mjeqn{\underset{p\times J}{\mathbf{B}}}{};
we recommend setting \code{d} to the smallest possible
value, which is
\mjeqn{J}{J};
the default value is \mjeqn{J}{J}.}

\item{Q}{The
\mjeqn{J\times J}{}
scale matrix hyper-prior for the inverse-Wishart-distributed
covariance matrix,
\mjeqn{\underset{J\times J}{\mathbf{\Sigma}}}{}.
Note that
\code{Q}
must be positive-definite.}

\item{nu_0}{The degrees of freedom hyper-prior for the inverse-Wishart-distributed
covariance matrix,
\mjeqn{\underset{q\times q}{\mathbf{G}}}{};
note that
\mjeqn{\underset{q\times q}{\mathbf{G}}}{}
represents the prior
covariance for each
of the
vectors of subject-specific (random effect) parameters,
\mjeqn{\underset{q\times 1}{\mathbf{b}_i}}{}.
We recommend setting \code{nu_0} to the smallest possible
value, which is
\mjeqn{q}{q};
the default value is \mjeqn{q}{q}.}

\item{C_0}{The
\mjeqn{q\times q}{}
scale matrix hyper-prior for the inverse-Wishart-distributed
covariance matrix,
\mjeqn{\underset{q\times q}{\mathbf{G}}}{}.
Note that
\code{C_0}
must be positive-definite.}

\item{alpha}{The shape hyper-prior for
the inverse-gamma-distributed
\mjeqn{\sigma^2}{};
note that \mjeqn{\sigma^2}{}
represents the prior
variance for
each vector of longitudinal outcome data,
of the longitudinal outcome data,
\mjeqn{\underset{J\times 1}{\mathbf{y}_i}}{}
for
\mjeqn{i = 1, \cdots, n}{}.
That is,
\mjeqn{\operatorname{cov}\left( \underset{J\times 1}{\mathbf{y}_i} \right) =
{{\sigma }^{2}}\underset{J\times J}{\mathop{{{\mathbf{I}}_{J}}}}\,}{}
\mjeqn{}{}.
The default value is \mjeqn{10^-3}{}.}

\item{gamma}{The scale hyper-prior for
the inverse-gamma-distributed
\mjeqn{\sigma^2}{}.
The default value is \mjeqn{10^-3}{}.}

\item{o}{The shape hyper-prior for
the inverse-gamma-distributed
\mjeqn{s^2}{};
note that \mjeqn{s^2}{}
represents the prior
variance
of the
spike-and-slab feature selection parameter,
\mjeqn{\tau_{l}^2}{}.
The default value is \mjeqn{10^-3}{}.}

\item{u}{The scale hyper-prior for
the inverse-gamma-distributed
\mjeqn{s^2}{}.
The default value is \mjeqn{10^-3}{}.}

\item{a_1}{The first shape hyper-prior for
the Beta-distributed
\mjeqn{{{\theta }_{{\tilde{\beta }}}}}{};
note that
\mjeqn{{{\theta }_{{\tilde{\beta }}}}}{}
represents the prior
probability
of the
group-level feature selection indicator variable,
\mjeqn{\pi_{0k}}{},
equaling
\mjeqn{1}{}.
The default value is \mjeqn{1}{1}.}

\item{b_1}{The second shape hyper-prior for
the Beta-distributed
\mjeqn{{{\theta }_{{\tilde{\beta }}}}}{}.
The default value is \mjeqn{1}{1}.}

\item{g}{The first shape hyper-prior for
the Beta-distributed
\mjeqn{{{\theta }_{{{\tau }^{2}}}}}{};
note that
\mjeqn{{{\theta }_{{{\tau }^{2}}}}}{}
represents the prior
probability
of the
individual feature selection indicator variable,
\mjeqn{\pi_{0l}}{},
equaling
\mjeqn{1}{}.
The default value is \mjeqn{1}{1}.}

\item{h}{The second shape hyper-prior for
the Beta-distributed
\mjeqn{{{\theta }_{{{\tau }^{2}}}}}{}.
The default value is \mjeqn{1}{}.}

\item{raw_MCMC_output}{If set to \code{TRUE}, the output for \code{BSSMfs} will include a data frame
containing un-summarized MCMC output.
The default value is \code{FALSE}.}
}
\value{
If \code{raw_MCMC_output=TRUE}, \code{BSSMfs} returns a list with two data frames named
\code{"MCMC_output"} and \code{"MCMC_summary"};
\code{"MCMC_output"} is a data frame with the un-summarized MCMC output, while
\code{"MCMC_summary"} is a data frame with the summarized MCMC output.
If \code{raw_MCMC_output=FALSE}, \code{BSSMfs} returns a list with only one data frame named
\code{"MCMC_summary"}

The \code{"MCMC_output"} data frame containing the un-summarized MCMC output has the following columns:
\itemize{
 \item{"feature"}{} A number denoting the \mjeqn{l=1,\cdots,p}{} feature parameters.

 \item{"occasion"}{} A character denoting the \mjeqn{j=1,\cdots,J}{} occasions; each occasion
 is denoted by the character \code{"tj"}.

 \item{"feature_group"}{} A number denoting the \mjeqn{k=1,\cdots,K}{} feature groups.

 \item{"subject"}{} A number denoting the \mjeqn{i=1,\cdots,n}{} subjects.

 \item{"parameter"}{} A character denoting the model parameter; the model parameters include:

 \itemize{

 \item{"b"}{} the subject-specific parameters,
 \mjeqn{\underset{q\times 1}{\mathop{{{\mathbf{b}}_{i}}}}\,}{}.

 \item{"beta"}{} The feature parameters,
 \mjeqn{\underset{p\times J}{\mathop{\mathbf{B}}}\,}{}.

 \item{"deviance"}{} The deviance of our model;
 \mjeqn{
 deviance=nJ\log \left( 2\pi {{\sigma }^{2}} \right)+
 \frac{1}{{{\sigma }^{2}}}
 {{\left(vec\left( \underset{J\times n}{\mathop{{{\mathbf{Y}}^{T}}}}\, \right) -
  \underset{nJ\times Jp}{\mathop{\mathbf{X}}}\,
 vec\left( \underset{p\times J}{\mathop{\mathbf{B}}}\, \right) -
 \underset{nJ\times nq}{\mathop{\mathbf{Z}}}\,
 \underset{nq\times 1}{\mathop{\mathbf{b}}}\, \right)}^{T}}
 \left(vec\left( \underset{J\times n}{\mathop{{{\mathbf{Y}}^{T}}}}\, \right)-
  \underset{nJ\times Jp}{\mathop{\mathbf{X}}}\,
 vec\left( \underset{p\times J}{\mathop{\mathbf{B}}}\, \right)-
 \underset{nJ\times nq}{\mathop{\mathbf{Z}}}\,
 \underset{nq\times 1}{\mathop{\mathbf{b}}}\, \right)+C
 }{},
 where
 \mjeqn{C}{}
 is a constant that cancels out upon comparing models.
 The deviance can be used to calculate the deviance information criteria
 (DIC).

 \item{"G"}{} The subject-specific parameter covariance matrix,
 \mjeqn{\underset{q\times q}{\mathop{\mathbf{G}}}\,}{}.

 \item{"MSE"}{} The mean squared error with respect to predicting the longitudinal
 outcome data,
 \mjeqn{\underset{J\times 1}{\mathbf{y}_i}}{}.
 Note that
 \mjeqn{
 \left( \underset{J\times J}{\mathop{{{\mathbf{I}}_{J}}}}\,
 \otimes \underset{1\times p}{\mathop{\mathbf{x}_{i}^{T}}}\, \right)
 vec\left( \underset{p\times J}{\mathop{\mathbf{B}}}\, \right)+
 \underset{J\times q}{\mathop{{{\mathbf{Z}}_{i}}}}\,
 \underset{q\times 1}{\mathop{{{\mathbf{b}}_{i}}}}\,}{}
 is used to predict
 \mjeqn{\underset{J\times 1}{\mathbf{y}_i}}{}.

 \item{"pi_0_k"}{} The group-level feature selection parameters,
 \mjeqn{{{\pi}_{0k}}}{}.

 \item{"pi_0_l"}{} The individual feature selection parameters,
 \mjeqn{{{\pi}_{0l}}}{}.

 \item{"s_2"}{} The variance parameter of
 \mjeqn{\tau _{l}^{2}}{},
 \mjeqn{s^{2}}{}.

 \item{"Sigma"}{} The feature parameter covariance matrix,
 \mjeqn{\underset{J\times J}{\mathop{\mathbf{\Sigma }}}\,}{}.

 \item{"sigma2"}{} The variance parameter of
 \mjeqn{\underset{J\times 1}{\mathbf{y}_i}}{},
 \mjeqn{\sigma^{2}}{}.

 \item{"tau_2_l"}{} The spike-and-slab parameters,
 \mjeqn{\tau _{l}^{2}}{}.

 \item{"theta_beta"}{} The probability parameter for group-level feature selection,
 \mjeqn{{{\theta }_{{\tilde{\beta }}}}}{}.

 \item{"theta_tau_2"}{} The probability parameter for individual feature selection,
 \mjeqn{{{\theta }_{\tau^{2}}}}{}.

 \item{"y"}{} The predicted values of
 \mjeqn{\underset{J\times 1}{\mathbf{y}_i}}{}.

 }


 \item{"row_occasion"}{} A character denoting the \mjeqn{J^2}{} elements of
 the feature parameter covariance matrix,
 \mjeqn{\underset{J\times J}{\mathop{\mathbf{\Sigma }}}\,}{};
 each element
 is denoted by the character \code{"tj"}.

 \item{"q"}{} A number denoting the \mjeqn{q}{} subject-specific/random effect parameters,
 \mjeqn{\underset{q\times 1}{\mathop{{{\mathbf{b}}_{i}}}}\,}{}.

 \item{"row_q"}{} A number denoting the \mjeqn{q^2}{} elements of the
 subject-specific parameter covariance matrix,
 \mjeqn{\underset{q\times q}{\mathop{\mathbf{G}}}\,}{}.

 \item{"comp_time"}{} The elapsed computation time for all \code{"nsim"}
 MCMC iterations to be completed.

 \item{"mcmc_iter"}{} The MCMC iteration number.

 \item{"value"}{} The numeric value of model parameter at each MCMC iteration.

}

The \code{"MCMC_summary"} data frame containg the summarized MCMC output has the following columns:
\itemize{

 \item{"feature"}{} A number denoting the \mjeqn{l=1,\cdots,p}{} feature parameters.

 \item{"occasion"}{} A character denoting the \mjeqn{j=1,\cdots,J}{} occasions; each occasion
 is denoted by the character string \code{"tj"}.

 \item{"feature_group"}{} A number denoting the \mjeqn{k=1,\cdots,K}{} feature groups.

 \item{"subject"}{} A number denoting the \mjeqn{i=1,\cdots,n}{} subjects.

 \item{"parameter"}{} A character denoting the model parameter; the model parameters are the
 same here as for the \code{"MCMC_output"} data frame.

 \item{"row_occasion"}{} A character denoting the \mjeqn{J^2}{} elements of
 the feature parameter covariance matrix,
 \mjeqn{\underset{J\times J}{\mathop{\mathbf{\Sigma }}}\,}{};
 each element
 is denoted by the character \code{"tj"}.

 \item{"q"}{} A number denoting the \mjeqn{q}{} subject-specific/random effect parameters.

 \item{"row_q"}{} A number denoting the \mjeqn{q^2}{} elements of the
 subject-specific parameter covariance matrix,
 \mjeqn{\underset{q\times q}{\mathop{\mathbf{G}}}\,}{}.

 \item{"MCMC_mean"}{} The numeric posterior mean for each model parameter.

 \item{"MCMC_median"}{} The numeric posterior median for each model parameter.

 \item{"MCMC_sd"}{} The numeric posterior standard deviation for each model parameter.

 \item{"MCMC_LCL"}{} The numeric posterior \mjeqn{2.5^{th}}{} percentile for each model parameter;
 used to construct the \mjeqn{95}{} percent credible interval (CI) for each model parameter.

 \item{"MCMC_UCL"}{} The numeric MCMC \mjeqn{97.5^{th}}{} percentile for each model parameter;
 used to construct the \mjeqn{95}{} percent CI for each model parameter.

 \item{"Geweke_diag"}{} The numeric Geweke MCMC diagnostic for each model parameter; calculated using
 the \code{coda} R package.

 \item{"lag_5"}{} The numeric lag 5 autocorrelation for each model parameter; calculated using
 the \code{coda} R package.

 \item{"lag_10"}{} The numeric lag 10 autocorrelation for each model parameter; calculated using
 the \code{coda} R package.

 \item{"ESS"}{} The numeric MCMC effective sample size for each model parameter; calculated using
 the \code{coda} R package.

 \item{"comp_time"}{} The elapsed computation time for all \code{"nsim"}
 MCMC iterations to be completed.



}
}
\description{
This function implements a Markov chain Monte Carlo (MCMC) Gibbs sampler to fit
a Bayesian specific-specific bi-level feature selection model;
this model is developed for the data scenario where we have time-invariant feature data and
irregularly-spaced longitudinal outcome data;
this model specifies time-varying
feature parameters, and allows for the incorporation of group structure into
the feature selection process.
\loadmathjax
}
\examples{

set.seed(123)

#Define data settings:

n<-100
p<-16
K<-4
J<-3
p_k<-rep(p/K,K)
q<-2

#----------------------------------------------------------------------------------#

#Generate time-invariant feature data:

X<-matrix(rnorm(n*p),nrow=n,ncol=p)

#----------------------------------------------------------------------------------#

#Define true B:

Beta_k<-list()

for(k in 1:K){

 Beta_k[[k]]<-matrix(rnorm(J*p_k[k]),nrow=p_k[k],ncol=J)

}

#Introduce sparsity:

#Group-level feature sparsity:

Beta_k[[1]]<-0*Beta_k[[1]]

#Individual feature sparsity:

Beta_k[[2]][1,]<-0*Beta_k[[2]][1,]

#Collapse:

Beta<-dplyr::bind_rows(lapply(Beta_k,as.data.frame))\%>\%as.matrix

#vec(Beta):

beta<-matrix(Beta,nrow=J*p,ncol=1)

#----------------------------------------------------------------------------------#

#Define subject-specific feature data (time-varying):

Z<-list()

for(i in 1:n){

  Z[[i]]<-matrix(0,nrow=J,ncol=q)

  #For random intercept:

  Z[[i]][,1]<-1

  #For random slope:

  Z[[i]][,2]<-sort(rexp(n=J,rate=1/20))

  #Make first occasion 0:

  Z[[i]][,2][1]<-0

}

#----------------------------------------------------------------------------------#

#Define sigma2:

sigma2<-1/20

sigma2_I_J<-sigma2*diag(J)

#Define G:

G<-matrix(c(1/2,-10^-3,
            -10^-3,10^-4),
            q,q,byrow = T)

#----------------------------------------------------------------------------------#

#Generate outcome data:

Y_transpose<-matrix(0,nrow=J,ncol=n)

for (i in 1:n){

  #This is X_i (J x Jp):

  X_kron_i<-diag(J)\%x\%t(X[i,])

  #Mean structure:

  SS_MS_i<-X_kron_i\%*\%beta

  #Covariance structure:

  cov_y_i<-Z[[i]]\%*\%G\%*\%t(Z[[i]])+sigma2_I_J

  #Outcome data:

  Y_transpose[,i]<-MASS::mvrnorm(n=1,
                     mu=SS_MS_i,
                     Sigma=cov_y_i)

}

#----------------------------------------------------------------------------------#

#Fit model to simulated data:

I_J<-diag(J)

I_q<-diag(q)

model_results<-

BSSMfs(

    Y_transpose,
    X,
    Z,
    p_k,

    nsim=2000,
    burn=1000,
    thin=1,

    Q=10^-3*I_J,

    C_0=10^-3*I_q
 )

#Inspect feature parameter estimates:

model_results$MCMC_summary\%>\%dplyr::filter(parameter=='beta')

#Inspect MSE of longitudinal outcome data:

model_results$MCMC_summary\%>\%dplyr::filter(parameter=='MSE')

#Calculate and inspect DIC, using the posterior variance of the
#deviance to calculate the effective number of model parameters:

model_results$MCMC_summary\%>\%
dplyr::filter(parameter=='deviance')\%>\%
dplyr::mutate(DIC=MCMC_mean+.5*(MCMC_sd^2))

}
\references{

}
